backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
asyncpg==0.29.0
psycopg2-binary==2.9.9
pydantic==2.5.0
pydantic-settings==2.1.0
pydantic[email]==2.5.0
celery==5.3.4
redis==5.0.1
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2
alpaca-py==0.16.0
python-dotenv==1.0.0
bcrypt==4.0.1
passlib==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
email-validator==2.1.0
tensorflow-cpu==2.17.1
RestrictedPython==6.2
